Add Bash testing scripts
Adds the files conduct_tests.sh, declare_vars.sh, and process_args.sh in the
testing directory.  The files by themselves are not stand-alone programs, but
are fragments to be sources into a testing script.

process_args.sh processes the command-line arguments passed to a testing script

declare_vars.sh declares variables in which to save testing data

conduct_tests.sh conducts the tests and writes appropriate output to stdout